oculus_user_management/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
colorlog==6.7.0
dataclasses-json==0.5.8
dict-to-protobuf==0.0.3.10
dnspython==2.3.0
facepy==1.0.12
future==0.18.3
grpcio==1.58.0
grpcio-tools==1.58.0
idna==3.4
marshmallow==3.19.0
marshmallow-enum==1.5.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
protobuf==4.24.3
pydantic==1.10.9
pymongo==4.3.3
python-dotenv==1.0.0
python-facebook-api==0.15.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.3
